Draw a playing field based on the tiles in a GameGrid
Images are loaded from a spritemap, which is an image containing every little image used by the app stitched together as to only load one resource.
When a tile is used you reference the image to use as a Rect (rectangle) which gives a frame from which to take pixel colors from.

To connect a tile with an image to render, a TileImageType is associated with each tile