Save and restore data on screen rotation
With our data structures made Parcelable, we can override onSaveInstanceState() and onRestoreInstanteState() on our Activity and in a simple way save and restore the values from memory.
To make this complete, the data must be fed into PlayingFieldView, so that a new grid will not be created by screen rotation, but rather resume using the old data.