Implemented game logic
This patch adds the business logic needed to handle player clicks on a tile.
The revealTile() method has a new signature that allows us to signal back to the View whether a mine was clicked on or not.
If a mine was clicked, GAME OVER is displayed. A new state variable is introduced in PlayingFieldView, to keep track if the game over screen is showing or not.
There are still a few glitches with the GAME OVER screen, as some or all tiles show as unrevealed, even though they are revealed